refactor: change nglo to serilog and add DI
