+++ maxwell_dmm/IDmmService.cs
@@ -1,0 +1,4 @@
+﻿public interface IDmmService
+{
+    void Run();
+}
