+++ maxwell_dmm/appsettings.json
@@ -1,0 +1,22 @@
+{
+  "Serilog": {
+    "Using": [ "Serilog.Sinks.Console", "Serilog.Sinks.File" ],
+    "MinimumLevel": "Information",
+    // Where do we want to write our logs to? Choose from a large number of sinks:
+    // https://github.com/serilog/serilog/wiki/Provided-Sinks.
+    "WriteTo": [
+      {
+        "Name": "Console"
+      },
+      {
+        "Name": "File",
+        "Args": {  "path": "Logs/log.txt", "rollingInterval": "Day" }
+      }
+    ],
+    "Enrich": [ "FromLogContext", "WithMachineName", "WithEnvironmentName", "WithThreadId", "WithSpan", "WithExceptionDetails"],
+    "Properties": {
+      "Application": "dmm-logger"
+    }
+  },
+  "AllowedHosts": "*"
+}
